metrics: OCP specific CA config to avoid extra config steps
On OCP, there are two kinds of namespaces: for system workloads, those
begin with either `kube-` or `openshift-` and for user workloads.
Because when SPO is distributed as an (optional) part of OpenShift, it
is a system workload, it is deployed into the
`openshift-security-profiles` namespace.

Those namespaces should be monitored by the built-in Prometheus instance
by default, without any extra steps and for that, one of the steps is
that the prometheus pod that is scraping the metrics must be able to
connect to the metrics endpoints with TLS.

This patch tries to guess if SPO is running as a system workload on OCP
by checking for the OCP CA type and checking if the namespace begins
with `openshift-`. In that case, it uses a well-known path where the
prometheus pod mounts the Service CA certficate as the CAFile parameter
in the ServiceMonitor.

When running on OCP, the service certificate is issued by the Service CA
as we can see:
```
oc get secrets metrics-server-cert -ojsonpath='{.data.tls\.crt}' | base64 -d | openssl x509 -text -noout | grep Issuer
        Issuer: CN = openshift-service-serving-signer@1670486466
```

And the file we put into the CAFile corresponds to a ConfigMap in the
`openshift-monitoring` namespace which is then mounted to that magic
path in the scraper pod. We can verify that the CM contains a CA cert
that signed our metrics service cert with:
```
oc get cm -nopenshift-monitoring openshift-service-ca.crt -ojsonpath='{.data.service-ca\.crt}' | openssl x509 -text -noout | grep Subject
        Subject: CN = openshift-service-serving-signer@1670486466
```