tree e0bac68fc9661a6dd0a9f7cae7f69d64c30ed44a parent 416a6b3a9b3dd3de84cf3bc084a3096f308c9ed4 author Jakub Hrozek <jhrozek@redhat.com> 1670503151 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1670588771 -0800
doc: Amend documentation to clarify that no action is needed to enable metrics on OCP

Because SPO is installed as a system workload on OCP, at least when
installing from the RH catalog, we don't have to enable monitoring of
user workloads to scrape metrics. In fact, it wouldn't have helped in
case of SPO being installed into a namespace that begins with
`openshift-`.