tree f51bb52538e8e1f0643707749ef060c1dd063bde parent 65c6bf4ee508cc9527cf4dfe92f5bd9e3e083485 author Jakub Hrozek <jhrozek@redhat.com> 1670934958 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1679084117 -0700
profilerecorder: gracefully skip containers that did not record anything

This is mostly concerning SELinux because even containers that don't
seemingly do anything at least record some syscalls, e.g. exec.

But with SELinux, if a container only performs actions that are allowed
by the container policy in the first place, no AVCs would be generated.
In that case, the call to get AVCs would return with a rather generic
error, which would then be retried by the profilerecording reconciler.
As a consequence, if a pod had multiple containers and the one with no
AVCs was collected before others, the profilerecorder would never reach
the other containers.

Instead, let's return an error matching the situation and handle the
absence of AVCs gracefully, by emptying whatever there might be in the
grpc cache and moving on to the next container.