Add `spoc`, the Security Profiles Operator CLI
We discussed in our last sync meeting that it would be helpful to have a
slimmer SPO CLI for faster execution in edge scenarios. For that we
would need a new binary, which can be used to directly interact with
features like the ebpf recorder.

Refers to https://github.com/kubernetes-sigs/security-profiles-operator/issues/1482

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>