tree 93edcaa6d335f32252750d853eec94b3ca3c74ee parent ee92f976cfe7f0438a2b727a4dd170da89aa0deb author Sascha Grunert <sgrunert@redhat.com> 1676901070 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1676916108 -0800
Add `spoc`, the Security Profiles Operator CLI

We discussed in our last sync meeting that it would be helpful to have a
slimmer SPO CLI for faster execution in edge scenarios. For that we
would need a new binary, which can be used to directly interact with
features like the ebpf recorder.

Refers to https://github.com/kubernetes-sigs/security-profiles-operator/issues/1482

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>