tree 7cedf62b7cede1e702cfc456b210524bbca757a3 parent e37aef54c8dd0067492659f012df0811b3114c03 author Jakub Hrozek <jhrozek@redhat.com> 1677071663 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1677139293 -0800
selinux: Allow using other container-selinux policy templates than container

The container-selinux package (at least on RHEL/Fedora) ships a number
of policy templates. One of them is called simply `container` is special
in the sense that using it also creates a `type process`, the
others do not.

The effect is that if the other templates such as `net_container` should
be used, the CIL policy must also include the raw container type.

This patch amends the generation of the CIL policy from the SPO-specific
format so that the `container` type is always used.