Update bundle CSV to match 0.5.2 tag
We want to ship 0.5.2 to include some updated dependencies, but we need
to make sure the CSV matches the tag so that the version we're trying to
build/install (0.5.2) is what we expect (not 0.5.0) and not what's in
the current CSV.

This step might have been missed in the last release. Adding it here so
things are up-to-date.