tree 972fc77d50ad82ad673031f99ea467ac761014f3 parent 98d0e094e59cfb9fe5400cc63512a06e76885dbe author Cosmin Cojocar <gcojocar@adobe.com> 1678792519 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1678867336 -0700
Fix memory optimization feature after upgrade to controller-runtime v0.14.5

It seems that BuilderWithOptins has introduced a bug when combining the options provided by the cluster
with the custom options, which breaks the memory optimization feature.

BuilderWithOptions creates a new scheme object which is configured in the cache object
but not in the cluster object where the original scheme object is maintained. This means that the
registered schemes in the cluster scheme are not reflected into the cache. This prevents the controller from
starting with error messages such as:

E0314 10:44:24.541264 1302480 source.go:146] controller-runtime/source "msg"="kind must be registered to the Scheme"
"error"="no kind is registered for the type v1beta1.SeccompProfile in scheme \"pkg/runtime/scheme.go:100\""