Verify that base profiles exist in GitHub container registry
We now feature two package repositories for the baseprofiles in
https://github.com/orgs/security-profiles/packages.

Means that we can use `spoc push` to manually maintain them for now. At
some later stage we can add automation to push them on the fly.

The repositories have to made public as well, otherwise `cosign verify`
will fail.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>