tree 68e1c4788bf7ca78284a9fc4af69ceb3d1631818 parent d3abeba58cbe11a22faca37fa0db210adacf89cf author Jakub Hrozek <jhrozek@redhat.com> 1680387459 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1680710744 -0700
Auto-select selinuxd image based on nodeInfo.OSImage

Cluster nodes might be upgraded between major OS releases where the
libselinux major version is incompatible. This means that the selinuxd
image must be based on images matching the OS release and the libselinux
major version.

But SPO must also be able to select the images based on what OS are the
cluster nodes running.

To implement that, this patch adds a new key to the
security-profiles-operator-profile configmap that maps regexes matching
nodeInfo.OSImage to selinuxd images. The operator then selects the right
selinuxd image, falling back to using the RELATED_IMAGE environment
variable.