tree 60c481c7bc59c63578851245d0cbc458e4a891c8 parent ad475b78207a80b8975f81f943d69a7307af3e8f author Jakub Hrozek <jhrozek@redhat.com> 1682064187 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1682323949 -0700
Make the selinuxd auto-selection logic use environment variables

This patch doesn't change the functionality in any way whatsoever, but
makes it easier for tools that reason about bundle or helm manifests
to do their job.

I hit such issue with a tool we use in OpenShift that was not able to detect
that I forgot to rewrite the quay.io upstream images in the configMap to
our downstream images.

We used to put the mapping between selinuxd images and OS directly into
a configmap. This is problematic for the aforementioned tools as a
container pullspec is hard to find in operator manifests. To make their
job easier, the configmap now points to an environment variable name
that is defined in the operator deployment manifest.