tree f32c4248c074310f539b9670013f19f4bc5d2a3a parent 7a969a23c85deec2efe425db7171fe962e8c82ce author Jakub Hrozek <jhrozek@redhat.com> 1685131263 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1685518905 -0700
Recorded profiles can optionally be disabled after creation

In many cases it's useful to just record profiles but use them at a
later stage, e.g. record in CI, deploy later by the end user.

This patch implements a new option `disableProfileAfterRecording` for
the `ProfileRecording` CR that creates the recorded profiles using a
new state `Disabled`. Such disabled profiles are not reconciled, but
can be merged if the profile recording is set to do that.

Even though this PR adds a new attribute to the CRs, the changes are
backwards compatible as the option is to disable the profile (not
explicitly enable) and the default, both on the CRD level and the Go
structure level is false.