Avoid listing pods in the recording webhook handler
Avoid listing pods in the recording webhook handler, because this operation is sometimes really slow leading to a webhook timeout.

It seems that listing pods in a cluster with many pods is quite slow. It can sometime takes over 30 seconds which is the maximum timeout for a webhook handler in Kubernetes.

Signed-off-by: Cosmin Cojocar <gcojocar@adobe.com>