tree c5f13a8e0af7456e7402084d9ee516a6994decf9 parent 604178eb394e4e7680b0c9b1ab75737edd3d2687 author Cosmin Cojocar <gcojocar@adobe.com> 1699015371 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1700131344 +0100
Avoid listing pods in the recording webhook handler

Avoid listing pods in the recording webhook handler, because this operation is sometimes really slow leading to a webhook timeout.

It seems that listing pods in a cluster with many pods is quite slow. It can sometime takes over 30 seconds which is the maximum timeout for a webhook handler in Kubernetes.

Signed-off-by: Cosmin Cojocar <gcojocar@adobe.com>