tree 733e6c35a8244e1b304e1a660ea54050367b1ac4 parent 5cbcd1d8ec65ca821874f67295a87ae58599d708 author Lance Bragstad <lbragsta@redhat.com> 1700582967 -0600 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1706019733 +0100
Move e2e manifest updates to the suite layer

Currently, e2e tests on an OpenShift cluster will hang because the
selinuxd container gets stuck in a crash loop. This is because the
manifest is updated to point to an container image that doesn't work for
OpenShift (it's not using the correct version of SELinuxd for the
version of RHCOS under OpenShift).

This commit moves the manifest updates to the suite layer, where it's
possible to detect if e2e tests are running vanilla kube, kind, or
OpenShift. If we're running on OpenShift, we just don't update selinuxd
container image locations since what is in the deploy/operator.yaml
works.