Fix namespace selector comparison for webhook
We should only compare supported labels and not everything on the
namespace, otherwise we can end up being stuck in a `SPOD` with the
`UPDATING` state. We now only compare supported webhook labels and use
them as indicator if it requires and update or not.

Fixes https://github.com/kubernetes-sigs/security-profiles-operator/issues/1981

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>