tree 24c02d473fb38c0e4fbb5744b0ff8fd5d2baa62d parent cdd5d7e6addfb4795f9404d21ebdfdd7dbedb755 author Sascha Grunert <sgrunert@redhat.com> 1701167287 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1701181192 +0100
Fix namespace selector comparison for webhook

We should only compare supported labels and not everything on the
namespace, otherwise we can end up being stuck in a `SPOD` with the
`UPDATING` state. We now only compare supported webhook labels and use
them as indicator if it requires and update or not.

Fixes https://github.com/kubernetes-sigs/security-profiles-operator/issues/1981

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>