tree 2f79f9ad66252e7c5b6165d4d1d64cc41e2daab7 parent 6181747147ac6427206c06bc4e37e8c7914641b6 author Cosmin Cojocar <ccojocar@google.com> 1733303809 +0000 committer Kubernetes Prow Robot <20407524+k8s-ci-robot@users.noreply.github.com> 1733319302 +0000
Add a ComplainMode flag into the apparmor profile

This falg will place the apparmor profile into compalin mode, otherwise
it defaults to enforce mode.

In complain mode, if a given action is not allowed, it will be allowed,
but this violation will be logged with a tag of access being "ALLOWED
unconfined".

Change-Id: I597883a528edf60f05f1cb7914429b3c5b05b2cd
Signed-off-by: Cosmin Cojocar <ccojocar@google.com>