tree 3fec69fd67a88c950d63d7beeda043632509c837 parent 264c1fdee3cf6e4cc3f0aa759af42d1283e15137 author Maximilian Hils <mhils@google.com> 1736949326 +0000 committer Kubernetes Prow Robot <20407524+k8s-ci-robot@users.noreply.github.com> 1737034716 -0800
bpf recorder: load bpf program on startup to win race with container startup

This commit makes it so that the BPF program is loaded (and relocated)
on profile recorder startup. When we then want to start the recording,
we only need to attach the programs. This fixes the flakiness in #2667
and #2668 where program loading is taking so long that we miss container
startup.