bpf recorder: load bpf program on startup to win race with container startup
This commit makes it so that the BPF program is loaded (and relocated)
on profile recorder startup. When we then want to start the recording,
we only need to attach the programs. This fixes the flakiness in #2667
and #2668 where program loading is taking so long that we miss container
startup.