can't load security-profiles-operator as dependency of another operator
When loaded through OperatorHub as a dependency of another operator,
security-profiles-operator is loaded in the namespace of the other operator.
When this happens, any created SelinuxProfile stays in the Pending State.

It appears that the function getDS() is searching for the DaemonSet with
a label of "spod". However, the r.client.List() call is returning all
DaemonSets in the Namespace and bailing because it found more than one.

This commit changes the logic to call Get() instead of GetList().

See: 2699

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>