after moving to cluster scoped, namespace still used
Recently SelinuxProfile was moved from Namespace scoped to Cluster
scoped. Some of the code was still using the Namespace as part of the
generation of the  "Usage" string. The Namespace was blank, so the usage
string just ended in an "_", something like "policyName_.process".
However, workloadannotator was using the application pods namespace and
the code strip the "_namspace.process" off the usage string to get the
actual profile name was failing. So just removed all the logic was
append "_namespace" to the profile name in the usage.

See: 2745

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>