CMP-3943, CMP-3941: Use ubi9-minimal and microdnf
We can use ubi9 minimal since it covers all our cases and is FIPS
validated. Also simplify the dnf and microdnf symlinking, which we don't
need to have a conditional for if we're always using a minimal image.
Using a minimal image is good since it reduces the overall surface area
of the container image.